requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
duckduckgo_search==8.0.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.1.0
faiss-cpu==1.11.0
fastapi==0.115.13
feedparser==6.0.11
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
gdown==5.2.0
google-auth==2.40.2
google-genai==1.17.0
googlesearch-python==1.3.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.0.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jellyfish==1.2.0
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
lightning-utilities==0.14.3
llvmlite==0.44.0
lxml==5.3.2
markdown-it-py==3.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.8.1
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
namex==0.0.9
narwhals==1.35.0
nest_asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
omegaconf==2.3.0
openpyxl==3.1.5
optree==0.15.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rdflib==7.1.4
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segtok==1.5.11
selenium==4.33.0
sentence-transformers==4.0.2
sentencepiece==0.2.0
sgmllib3k==1.0.0
shap==0.46.0
sklearn-compat==0.1.3
slicer==0.0.8
smote==0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SPARQLWrapper==2.0.0
stack_data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics==1.7.1
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
tzdata==2025.2
ucimlrepo==0.0.7
undetected-chromedriver==3.5.5
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xgboost==3.0.0
xxhash==3.5.0
yake==0.4.8
yarl==1.19.0
